requirements.txt
Bootstrap_Flask==2.2.0
Bootstrap_Flask==2.4.1
Flask==3.1.0
numpy==2.2.3
Pillow==11.1.0
Werkzeug==3.1.3
